Moved get_new_topic_id to the service file.
