+++ core/controllers/skill_editor_test.py
@@ -16,7 +16,7 @@
 
 from core.domain import role_services
 from core.domain import skill_services
-from core.domain import topic_fetchers
+from core.domain import topic_services
 from core.domain import user_services
 from core.platform import models
 from core.tests import test_utils
@@ -45,7 +45,7 @@ class BaseSkillEditorControllerTests(test_utils.GenericTestBase):
         self.save_new_skill(self.skill_id, self.admin_id, 'Description')
         self.skill_id_2 = skill_services.get_new_skill_id()
         self.save_new_skill(self.skill_id_2, self.admin_id, 'Description')
-        self.topic_id = topic_fetchers.get_new_topic_id()
+        self.topic_id = topic_services.get_new_topic_id()
         self.save_new_topic(
             self.topic_id, self.admin_id, 'Name', 'Description',
             [], [], [self.skill_id], [], 1)
