+++ core/domain/topic_fetchers_test.py
@@ -40,7 +40,7 @@ class TopicFetchersUnitTests(test_utils.GenericTestBase):
 
     def setUp(self):
         super(TopicFetchersUnitTests, self).setUp()
-        self.TOPIC_ID = topic_fetchers.get_new_topic_id()
+        self.TOPIC_ID = topic_services.get_new_topic_id()
         changelist = [topic_domain.TopicChange({
             'cmd': topic_domain.CMD_ADD_SUBTOPIC,
             'title': 'Title',
@@ -67,12 +67,6 @@ class TopicFetchersUnitTests(test_utils.GenericTestBase):
         self.user_a = user_services.UserActionsInfo(self.user_id_a)
         self.user_b = user_services.UserActionsInfo(self.user_id_b)
         self.user_admin = user_services.UserActionsInfo(self.user_id_admin)
-
-    def test_get_new_topic_id(self):
-        new_topic_id = topic_fetchers.get_new_topic_id()
-
-        self.assertEqual(len(new_topic_id), 12)
-        self.assertEqual(topic_models.TopicModel.get_by_id(new_topic_id), None)
 
     def test_get_topic_from_model(self):
         topic_model = topic_models.TopicModel.get(self.TOPIC_ID)
@@ -115,7 +109,7 @@ class TopicFetchersUnitTests(test_utils.GenericTestBase):
         self.assertEqual(topic.to_dict(), expected_topic)
 
     def test_get_topic_by_version(self):
-        topic_id = topic_fetchers.get_new_topic_id()
+        topic_id = topic_services.get_new_topic_id()
         self.save_new_topic(
             topic_id, self.user_id, 'topic name', 'Description',
             [], [], [], [], 1)
